# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_server_endpoints.AgentLogSummary:
    properties:
      agent_type:
        type: string
      completed_at:
        type: string
      error:
        type: string
      id:
        type: string
      iterations:
        type: integer
      started_at:
        type: string
      success:
        type: boolean
    type: object
...
  internal_server_endpoints.BookMetadata:
    properties:
      author:
        type: string
      authors:
        items:
          type: string
        type: array
      cover_page:
        type: integer
      description:
        type: string
      isbn:
        type: string
      language:
        type: string
      lccn:
        type: string
      publication_year:
        type: integer
      publisher:
        type: string
      subjects:
        items:
          type: string
        type: array
      subtitle:
        type: string
      title:
        type: string
    type: object
...
  internal_server_endpoints.DetailedJobStatusResponse:
    properties:
      agent_logs:
        description: Agent logs summary
        items:
          $ref: '#/definitions/internal_server_endpoints.AgentLogSummary'
        type: array
      book_id:
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/internal_server_endpoints.MetadataStatus'
        description: Metadata status
      ocr_progress:
        additionalProperties:
          $ref: '#/definitions/internal_server_endpoints.ProviderProgress'
        description: Per-provider OCR progress
        type: object
      stages:
        allOf:
        - $ref: '#/definitions/internal_server_endpoints.StageProgress'
        description: Stage progress with costs
      structure:
        allOf:
        - $ref: '#/definitions/internal_server_endpoints.StructureStatus'
        description: Structure status (common-structure job)
      toc:
        allOf:
        - $ref: '#/definitions/internal_server_endpoints.ToCStatus'
        description: ToC status
      total_pages:
        type: integer
    type: object
  internal_server_endpoints.DiscoveredPattern:
    properties:
      heading_format:
        type: string
      level:
        type: integer
      level_name:
        type: string
      pattern_type:
        type: string
      range_end:
        type: string
      range_start:
        type: string
      reasoning:
        type: string
    type: object
...
  internal_server_endpoints.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_server_endpoints.ExcludedRange:
    properties:
      end_page:
        type: integer
      reason:
        type: string
      start_page:
        type: integer
    type: object
...
  internal_server_endpoints.MetadataStatus:
    properties:
      complete:
        type: boolean
      cost_usd:
        type: number
      data:
        $ref: '#/definitions/internal_server_endpoints.BookMetadata'
      failed:
        type: boolean
      started:
        type: boolean
    type: object
...
  internal_server_endpoints.PatternAnalysisResult:
    properties:
      excluded_ranges:
        items:
          $ref: '#/definitions/internal_server_endpoints.ExcludedRange'
        type: array
      patterns:
        items:
          $ref: '#/definitions/internal_server_endpoints.DiscoveredPattern'
        type: array
      reasoning:
        type: string
    type: object
...
  internal_server_endpoints.ProviderProgress:
    properties:
      complete:
        type: integer
      cost_usd:
        type: number
      total:
        type: integer
    type: object
...
  internal_server_endpoints.StageProgress:
    properties:
      ocr:
        properties:
          complete:
            type: integer
          cost_by_provider:
            additionalProperties:
              format: float64
              type: number
            type: object
          total:
            type: integer
          total_cost_usd:
            type: number
        type: object
      pattern_analysis:
        properties:
          complete:
            type: boolean
          cost_usd:
            type: number
        type: object
    type: object
...
  internal_server_endpoints.StructureStatus:
    properties:
      chapter_count:
        type: integer
      chapters_extracted:
        type: integer
      chapters_polished:
        type: integer
      chapters_total:
        type: integer
      complete:
        type: boolean
      cost_usd:
        type: number
      failed:
        type: boolean
      phase:
        description: Phase tracking (build -> extract -> classify -> polish -> finalize)
        type: string
      polish_failed:
        type: integer
      retries:
        type: integer
      started:
        type: boolean
    type: object
...
  internal_server_endpoints.ToCEntry:
    properties:
      actual_page_num:
        type: integer
      entry_number:
        type: string
      is_linked:
        type: boolean
      level:
        type: integer
      level_name:
        type: string
      printed_page_number:
        type: string
      sort_order:
        type: integer
      source:
        description: '"extracted" or "discovered"'
        type: string
      title:
        type: string
    type: object
  internal_server_endpoints.ToCStatus:
    properties:
      cost_usd:
        type: number
      discover_complete:
        description: All entries discovered
        type: boolean
      end_page:
        type: integer
      entries:
        items:
          $ref: '#/definitions/internal_server_endpoints.ToCEntry'
        type: array
      entries_discovered:
        description: Actually discovered (source="discovered")
        type: integer
      entries_linked:
        type: integer
      entries_to_find:
        description: From pattern analysis (how many should be discovered)
        type: integer
      entry_count:
        description: Entries (when extracted)
        type: integer
      excluded_ranges:
        description: Number of excluded page ranges
        type: integer
      extract_complete:
        type: boolean
      extract_failed:
        type: boolean
      extract_started:
        description: Extract stage
        type: boolean
      finalize_complete:
        type: boolean
      finalize_failed:
        type: boolean
      finalize_retries:
        type: integer
      finalize_started:
        description: Finalize stage (overall)
        type: boolean
      finder_complete:
        type: boolean
      finder_failed:
        type: boolean
      finder_started:
        description: Finder stage
        type: boolean
      found:
        type: boolean
      link_complete:
        type: boolean
      link_failed:
        type: boolean
      link_retries:
        type: integer
      link_started:
        description: Link stage
        type: boolean
      pattern_analysis:
        allOf:
        - $ref: '#/definitions/internal_server_endpoints.PatternAnalysisResult'
        description: Full pattern analysis result
      pattern_complete:
        description: 'Finalize sub-phases: Pattern Analysis → Chapter Discovery →
          Gap Validation'
        type: boolean
      patterns_found:
        description: Number of patterns discovered
        type: integer
      start_page:
        type: integer
      validate_complete:
        description: Gap validation done (same as FinalizeComplete for now)
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/jackzampolin/shelf
  description: Book digitization pipeline API for managing books, jobs, and processing.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Shelf API
  version: "1.0"
paths:
...
  /api/jobs/status/{book_id}/detailed:
    get:
      description: Get comprehensive processing status including per-provider OCR
        progress, costs, metadata, and ToC details
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.DetailedJobStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Get detailed job status for a book
      tags:
      - jobs
...
swagger: "2.0"